Update wpa_supplicant to 2.6, patch upstream CVEs
This update addresses the so-called "Krack Attack" on wpa_supplicant:

 - CVE-2017-13077
 - CVE-2017-13078
 - CVE-2017-13079
 - CVE-2017-13080
 - CVE-2017-13081
 - CVE-2017-13082
 - CVE-2017-13086
 - CVE-2017-13087
 - CVE-2017-13088

Credit to @kyentei for his timely report to Solus so that it might be dealt
with swiftly. Additionally, kudos to the original security company for
discovering these WPA2 weaknesses. For full details, please ensure to visit
their [website](https://www.krackattacks.com/).

This resolves T4785 within Solus.

Maniphest Tasks: T4785

Signed-off-by: Ikey Doherty <ikey@solus-project.com>